requirements.txt
Django==3.2.9
requests==2.26.0
djangorestframework==3.12.4
pandas==1.3.4
drf-yasg==1.20.0